Add scad-mode and scad-preview
This also adds some code that tries to search every scad buffer for
simple assignments of $vpr, $vpt and $vpd to determine the viewport of
the scad. This viewport is then written to the scad-prview default
viewport. When scad-preview ist opened or whenever the viewport in the
scad-preview buffer is reset, the viewport of the scad drawing is used
as the default rather than the (0 0 0 50 0 20 500) default of scad-preview.